new version

+++ Classes/Controller/FolgekontaktController.php
@@ -1,15 +1,17 @@
 <?php
 namespace Ud\Iqtp13db\Controller;
+use \Datetime;
+use TYPO3\CMS\Extbase\Utility\DebuggerUtility;
 
 /***
  *
- * This file is part of the "IQ Webapp Anerkennungsberatung" Extension for TYPO3 CMS.
+ * This file is part of the "IQ Webapp Anerkennungserstberatung" Extension for TYPO3 CMS.
  *
  * For the full copyright and license information, please read the
  * LICENSE.txt file that was distributed with this source code.
  *
- *  (c) 2020 Uli Dohmen <edv@whkt.de>, WHKT
- * 
+ *  (c) 2022 Uli Dohmen <edv@whkt.de>, WHKT
+ *
  ***/
 
 /**
@@ -18,29 +20,23 @@ namespace Ud\Iqtp13db\Controller;
 class FolgekontaktController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
 {
    
+    protected $niqinterface, $niqbid, $usergroup;
+    
     /**
      * folgekontaktRepository
      *
      * @var \Ud\Iqtp13db\Domain\Repository\FolgekontaktRepository
      * @TYPO3\CMS\Extbase\Annotation\Inject
      */
-    protected $folgekontaktRepository = NULL;
+    protected $folgekontaktRepository;
     
-    /**
-     * beratungRepository
-     *
-     * @var \Ud\Iqtp13db\Domain\Repository\BeratungRepository
-     * @TYPO3\CMS\Extbase\Annotation\Inject
-     */
-    protected $beratungRepository = NULL;
-
     /**
      * teilnehmerRepository
      *
      * @var \Ud\Iqtp13db\Domain\Repository\TeilnehmerRepository
      * @TYPO3\CMS\Extbase\Annotation\Inject
      */
-    protected $teilnehmerRepository = NULL;
+    protected $teilnehmerRepository;
 
     /**
      * beraterRepository
@@ -48,8 +44,48 @@ class FolgekontaktController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCon
      * @var \Ud\Iqtp13db\Domain\Repository\BeraterRepository
      * @TYPO3\CMS\Extbase\Annotation\Inject
      */
-    protected $beraterRepository = NULL;      
+    protected $beraterRepository;      
         
+    /**
+     * abschlussRepository
+     *
+     * @var \Ud\Iqtp13db\Domain\Repository\AbschlussRepository
+     * @TYPO3\CMS\Extbase\Annotation\Inject
+     */
+    protected $abschlussRepository;	
+    
+    /**
+     * frontendUserGroupRepository
+     *
+     * @var \TYPO3\CMS\Extbase\Domain\Repository\FrontendUserGroupRepository
+     * @TYPO3\CMS\Extbase\Annotation\Inject
+     */
+    protected $frontendUserGroupRepository;
+    
+    /**
+     * action init
+     *
+     * @param void
+     */
+    public function initializeAction()
+    {
+        $this->user=null;
+        $context = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(\TYPO3\CMS\Core\Context\Context::class);
+        if($context->getPropertyFromAspect('frontend.user', 'isLoggedIn')){
+            $this->user=$GLOBALS['TSFE']->fe_user->user;
+        }
+        
+        //DebuggerUtility::var_dump($this->user);
+        if($this->user != NULL) {
+            $standardniqidberatungsstelle = $this->settings['standardniqidberatungsstelle'];
+            $this->usergroup = $this->frontendUserGroupRepository->findByIdentifier($this->user['usergroup']);
+            $userniqidbstelle = $this->usergroup->getDescription();
+            $this->niqbid = $userniqidbstelle == '' ? $standardniqidberatungsstelle : $userniqidbstelle;
+        }
+                
+        $this->niqinterface = new \Ud\Iqtp13db\Helper\NiqInterface();
+    }
+    
     /**
      * action show
      *
@@ -69,27 +105,28 @@ class FolgekontaktController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCon
     /**
      * action new
      *
-     * @param \Ud\Iqtp13db\Domain\Model\Beratung $beratung
+     * @param \Ud\Iqtp13db\Domain\Model\Teilnehmer $teilnehmer
      * @return void
      */
-    public function newAction(\Ud\Iqtp13db\Domain\Model\Beratung $beratung)
+    public function newAction(\Ud\Iqtp13db\Domain\Model\Teilnehmer $teilnehmer)
     {
-    	$username = $GLOBALS['TSFE']->fe_user->user['username'];
-    	
-    	$berater = $this->beraterRepository->findOneByKuerzel($username);
-    	if($berater == NULL) {
-    		$berater = $this->beraterRepository->findOneByKuerzel('admin');
-    	} 
-    	$teilnehmer = $beratung->getTeilnehmer();
-    	    	
-    	$alleberater = $this->beraterRepository->findAll();
-    	$this->view->assign('alleberater', $alleberater);
-    	$this->view->assign('berater', $berater);
-        $this->view->assign('teilnehmer', $teilnehmer);
-        
         $valArray = $this->request->getArguments();
-        $this->view->assign('calleraction', $valArray['calleraction']);
-        $this->view->assign('callercontroller', $valArray['callercontroller']);
+        
+        if($teilnehmer->getNiqchiffre() != '' && $this->niqinterface->check_curl() == FALSE) {
+            $this->addFlashMessage("Bearbeiten von bereits übertragenen Datensätzen vorübergehend nicht möglich, da NIQ-Datenbank nicht erreichbar.", '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, array('callerpage' => $valArray['callerpage']));
+        } else {
+            $alleberater = $this->beraterRepository->findBerater4Group($this->settings['beraterstoragepid'], $this->user['usergroup']);
+        	
+        	$this->view->assign('alleberater', $alleberater);
+        	$this->view->assign('berater', $this->user);
+            $this->view->assign('teilnehmer', $teilnehmer);
+            
+            $this->view->assign('calleraction', $valArray['calleraction']);
+            $this->view->assign('callercontroller', $valArray['callercontroller']);
+            $this->view->assign('callerpage', $valArray['callerpage']);
+            $this->view->assign('settings', $this->settings);
+        }
     }
     
 
@@ -101,18 +138,63 @@ class FolgekontaktController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCon
      */
     public function createAction(\Ud\Iqtp13db\Domain\Model\Folgekontakt $folgekontakt)
     {  		
-        //$this->addFlashMessage('Folgekontakt wurde erstellt.', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
-        $this->folgekontaktRepository->add($folgekontakt);
-        
-        // Daten sofort in die Datenbank schreiben
-        $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
-        $persistenceManager->persistAll();
-                
-        $this->view->assign('calleraction', 'listarchiv');
-        $this->view->assign('callercontroller', 'Beratung');
-        
         $valArray = $this->request->getArguments();
-        $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, null);
+        $teilnehmer = $this->teilnehmerRepository->findByUid($valArray['folgekontakt']['teilnehmer']);
+        //DebuggerUtility::var_dump($teilnehmer);
+        //die;
+        
+        if($teilnehmer->getNiqchiffre() != '' && $this->niqinterface->check_curl() == FALSE) {
+            $this->addFlashMessage("Bearbeiten von bereits übertragenen Datensätzen vorübergehend nicht möglich, da NIQ-Datenbank nicht erreichbar.", '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, array('callerpage' => $valArray['callerpage']));
+        } else {
+            $teilnehmer = $folgekontakt->getTeilnehmer();
+            
+            $letzterfolgekontakt = $this->folgekontaktRepository->findLastByTNuid($teilnehmer->getUid());
+            
+            $beratungtimestamp = DateTime::createFromFormat("Y-m-d", $teilnehmer->getBeratungdatum());
+            $folgekontakttimestamp = DateTime::createFromFormat("d.m.Y", $folgekontakt->getDatum());  
+            $letzterfolgekontakttimestamp = $letzterfolgekontakt != NULL ? DateTime::createFromFormat("d.m.Y", $letzterfolgekontakt->getDatum()) : DateTime::createFromFormat("d.m.Y", '01.01.1970');
+            
+            if($folgekontakttimestamp->getTimestamp() <= $beratungtimestamp->getTimestamp() || $folgekontakttimestamp->getTimestamp() <= $letzterfolgekontakttimestamp->getTimestamp()) {
+                $this->addFlashMessage('Datum Folgekontakt muss nach letztem Folgekontakt und Datum Erstberatung sein.', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+                
+                $this->redirect('new', 'Folgekontakt', null, array('teilnehmer' => $teilnehmer, 'calleraction' => $valArray['calleraction'],'callercontroller' => $valArray['callercontroller'], 'callerpage' => $valArray['callerpage']));
+            } else {
+                $this->folgekontaktRepository->add($folgekontakt);
+                // Daten sofort in die Datenbank schreiben
+                $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
+                $persistenceManager->persistAll();
+                
+                if($teilnehmer->getNiqchiffre() != '') {
+                    $abschluesse = $this->abschlussRepository->findByTeilnehmer($teilnehmer);
+                    $folgekontakte = $this->folgekontaktRepository->findByTeilnehmer($teilnehmer);                
+                    if($teilnehmer->getNiqidberatungsstelle() != '0') {
+                        $niqidbstelle = $teilnehmer->getNiqidberatungsstelle();
+                    } else {
+                        $niqidbstelle = $this->niqbid;
+                    }
+                    $returnarray = $this->niqinterface->uploadtoNIQ($teilnehmer, $abschluesse, $folgekontakte, $niqidbstelle);
+                    
+//                    DebuggerUtility::var_dump($returnarray[0]);
+ //                   die;
+                    $retteilnehmer = $this->teilnehmerRepository->findByUid($returnarray[0]->getUid());
+                    $retstring = $returnarray[1];
+                    
+                    $this->teilnehmerRepository->update($retteilnehmer);
+                    
+                    $this->addFlashMessage($retstring, '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
+                }
+                
+                // Daten sofort in die Datenbank schreiben
+                $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
+                $persistenceManager->persistAll();
+                
+                $this->view->assign('calleraction', 'listarchiv');
+                $this->view->assign('callercontroller', 'Teilnehmer');
+            }
+               
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, null);
+        }
     }
 	
 	 /**
@@ -124,15 +206,25 @@ class FolgekontaktController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCon
      */
     public function editAction(\Ud\Iqtp13db\Domain\Model\Folgekontakt $folgekontakt)
     {
-   		$berater = $this->beraterRepository->findOneByKuerzel('admin');
-   		$this->view->assign('berater', $berater);
-    	$alleberater = $this->beraterRepository->findAll();    	
-    	$this->view->assign('alleberater', $alleberater);
-        $this->view->assign('folgekontakt', $folgekontakt);
-        
         $valArray = $this->request->getArguments();
-        $this->view->assign('calleraction', $valArray['calleraction']);
-        $this->view->assign('callercontroller', $valArray['callercontroller']);
+        $teilnehmer = $folgekontakt->getTeilnehmer();
+        
+        if($teilnehmer->getNiqchiffre() != '' && $this->niqinterface->check_curl() == FALSE) {
+            $this->addFlashMessage("Bearbeiten von bereits übertragenen Datensätzen vorübergehend nicht möglich, da NIQ-Datenbank nicht erreichbar.", '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, array('callerpage' => $valArray['callerpage']));
+        } else {
+       		$this->view->assign('berater', $this->user);
+       		$alleberater = $this->beraterRepository->findBerater4Group($this->settings['beraterstoragepid'], $this->user['usergroup']);
+       		
+        	$this->view->assign('alleberater', $alleberater);
+            $this->view->assign('folgekontakt', $folgekontakt);
+            $this->view->assign('teilnehmer', $teilnehmer);
+            
+            $this->view->assign('callerpage', $valArray['callerpage']);
+            $this->view->assign('calleraction', $valArray['calleraction']);
+            $this->view->assign('callercontroller', $valArray['callercontroller']);
+            $this->view->assign('settings', $this->settings);
+        }
     }
 
     /**
@@ -143,16 +235,45 @@ class FolgekontaktController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCon
      */
     public function updateAction(\Ud\Iqtp13db\Domain\Model\Folgekontakt $folgekontakt)
     {
-    	
-        //$this->addFlashMessage('Folgekontakt aktualisiert.', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
-        $this->folgekontaktRepository->update($folgekontakt);
-         
-        // Daten sofort in die Datenbank schreiben
-        $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
-        $persistenceManager->persistAll();
-        
         $valArray = $this->request->getArguments();
-        $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, null);
+        $teilnehmer = $folgekontakt->getTeilnehmer();
+        
+        if($teilnehmer->getNiqchiffre() != '' && $this->niqinterface->check_curl() == FALSE) {
+            $this->addFlashMessage("Bearbeiten von bereits übertragenen Datensätzen vorübergehend nicht möglich, da NIQ-Datenbank nicht erreichbar.", '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, array('callerpage' => $valArray['callerpage']));
+        } else {
+            $this->folgekontaktRepository->update($folgekontakt);
+             
+            // Daten sofort in die Datenbank schreiben
+            $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
+            $persistenceManager->persistAll();
+            
+            $teilnehmer = $folgekontakt->getTeilnehmer();
+            
+            if($teilnehmer->getNiqchiffre() != '') {
+                $abschluesse = $this->abschlussRepository->findByTeilnehmer($teilnehmer);
+                $folgekontakte = $this->folgekontaktRepository->findByTeilnehmer($teilnehmer);
+                if($teilnehmer->getNiqidberatungsstelle() != '0') {
+                    $niqidbstelle = $teilnehmer->getNiqidberatungsstelle();
+                } else {
+                    $niqidbstelle = $this->niqbid;
+                }
+                $returnarray = $this->niqinterface->uploadtoNIQ($teilnehmer, $abschluesse, $folgekontakte, $niqidbstelle);
+                
+                $retteilnehmer = $this->teilnehmerRepository->findByUid($returnarray[0]->getUid());
+                $retstring = $returnarray[1];
+                
+                $this->teilnehmerRepository->update($retteilnehmer);
+                
+                $this->addFlashMessage($retstring, '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
+            }
+            
+            // Daten sofort in die Datenbank schreiben
+            $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
+            $persistenceManager->persistAll();
+            
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, array('callerpage' => $valArray['callerpage']));
+        }
     }
 
     /**
@@ -163,19 +284,43 @@ class FolgekontaktController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCon
      */
     public function deleteAction(\Ud\Iqtp13db\Domain\Model\Folgekontakt $folgekontakt)
     {
-        //$this->addFlashMessage('Folgekontakt gelöscht.', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
-        $this->folgekontaktRepository->remove($folgekontakt);
-        
-        // Daten sofort in die Datenbank schreiben
-        $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
-        $persistenceManager->persistAll();
-        
         $valArray = $this->request->getArguments();
-        $this->view->assign('calleraction', $valArray['calleraction']);
-        $this->view->assign('callercontroller', $valArray['callercontroller']);
+        $teilnehmer = $folgekontakt->getTeilnehmer();
         
-        $valArray = $this->request->getArguments();
-        $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, null);
+        if($teilnehmer->getNiqchiffre() != '' && $this->niqinterface->check_curl() == FALSE) {
+            $this->addFlashMessage("Bearbeiten von bereits übertragenen Datensätzen vorübergehend nicht möglich, da NIQ-Datenbank nicht erreichbar.", '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, array('callerpage' => $valArray['callerpage']));
+        } else {
+            $this->folgekontaktRepository->remove($folgekontakt);
+            
+            // Daten sofort in die Datenbank schreiben
+            $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
+            $persistenceManager->persistAll();
+            
+            $teilnehmer = $folgekontakt->getTeilnehmer();
+            if($teilnehmer->getNiqchiffre() != '') {
+                $abschluesse = $this->abschlussRepository->findByTeilnehmer($teilnehmer);
+                $folgekontakte = $this->folgekontaktRepository->findByTeilnehmer($teilnehmer);
+                if($teilnehmer->getNiqidberatungsstelle() != '0') {
+                    $niqidbstelle = $teilnehmer->getNiqidberatungsstelle();
+                } else {
+                    $niqidbstelle = $this->niqbid;
+                }
+                $returnarray = $this->niqinterface->uploadtoNIQ($teilnehmer, $abschluesse, $folgekontakte, $niqidbstelle);
+                
+                $retteilnehmer = $returnarray[0];
+                $retstring = $returnarray[1];
+                
+                $this->teilnehmerRepository->update($retteilnehmer);
+                
+                $this->addFlashMessage($retstring, '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
+            }
+            // Daten sofort in die Datenbank schreiben
+            $persistenceManager = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Extbase\\Persistence\\Generic\\PersistenceManager');
+            $persistenceManager->persistAll();
+
+            $this->redirect($valArray['calleraction'], $valArray['callercontroller'], null, array('callerpage' => $valArray['callerpage']));
+        }
         
     }
